+++ R/predictEnmSdm.r
@@ -69,10 +69,17 @@ predictEnmSdm <- function(
 	} else if (cores > 1L & nrows > n) {
 
 		`%makeWork%` <- foreach::`%dopar%`
-		cl <- parallel::makeCluster(cores, setup_strategy = 'sequential')
+		# cl <- parallel::makeCluster(cores, setup_strategy = 'sequential')
+		cl <- parallel::makeCluster(cores)
+		parallel::clusterEvalQ(cl, requireNamespace('parallel', quietly=TRUE))
 		doParallel::registerDoParallel(cl)
 		on.exit(parallel::stopCluster(cl), add=TRUE)
 			
+		# `%makeWork%` <- doRNG::`%dorng%`
+		# doFuture::registerDoFuture()
+		# future::plan(future::multisession(workers = cores))
+		# on.exit(future:::ClusterRegistry('stop'), add=TRUE)
+
 		paths <- .libPaths() # need to pass this to avoid "object '.doSnowGlobals' not found" error!!!
 		mcOptions <- list(preschedule = TRUE, set.seed = TRUE, silent = TRUE)